+++ pages/_app.js
@@ -1,7 +1,6 @@
 import "../styles/globals.css";
-import "../styles/flex2html.css"
-import { LiffProvider } from '../context/LiffContext';
-// import { initLiff } from "../api/line/liff";
+import "../styles/flex2html.css";
+import { LiffProvider } from '../hooks/useLiff';
 import { useState, useEffect } from "react";
 
 function MyApp({ Component, pageProps }) {
