+++ pages/home.js
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react";
-import { useLiff } from "../context/LiffContext";
 export default function Home() {  
   return (
     <div>
